pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "starplot"
authors = [{name = "Steve Berardi", email = "hello@steveberardi.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Matplotlib",
]
dynamic = ["version", "description"]
keywords = ["astronomy", "stars", "charts", "maps", "constellations", "sky", "plotting"]
dependencies = [
    "matplotlib >= 3.8.0",
    "numpy >= 1.26.2",
    "pandas >= 1.4.0",
    "pydantic >= 2.10.6",
    "shapely >= 2.0.1",
    "skyfield >= 1.41",
    "cartopy >= 0.21.1",
    "geopandas >= 1.0.1",
    "pillow >= 10.0.0",
    "PyYAML >= 6.0.1",
    "pyarrow >= 14.0.2",
    "pyogrio >= 0.10.0",
    "rtree >= 1.2.0",
    "requests >= 2.31.0",
    "duckdb >= 1.4.4",
    "sqlglot >= 28.9.0",
    "ibis-framework[duckdb,geospatial] >= 11",
    "astropy-healpix >= 1.1.2"
]

[project.optional-dependencies]
interactive = ["plotly>=5.0", "kaleido>=0.2"]

[project.urls]
Home = "https://starplot.dev"
Documentation = "https://starplot.dev"
Source = "https://github.com/steveberardi/starplot"

[project.scripts]
starplot = "starplot.cli:main"

[tool.flit.sdist]
include = [
    "src/starplot/data/library/*.parquet",
]

[tool.pytest.ini_options]
testpaths = ["./tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff.lint]
ignore = ["E501"]

[tool.marimo.display]
theme = "dark"
